Create README.md
This project is a player statistics website for the video game League of Legends. It is currently a work in. progress so the developer API key needs to be re-generated every few days for the website to work. 
Google Charts, Html, Python, CSS. JavaScript, Riot API, and Cassiopeia are some of the tools used to build this website.
The website can be accessed through the following link:
https://mushiewaffle.github.io/RivaLoL/